jobmaster/worker/requirements.txt
celery
redis
python-dotenv
psycopg2-binary
numpy
pandas
scikit-learn